backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.3.0
python-multipart==0.0.9
aiofiles==23.2.1
python-dotenv==1.0.1
chromadb==0.5.0
anthropic==0.28.0
langchain==0.2.5
langchain-anthropic==0.1.15
langchain-community==0.2.5
langchain-chroma==0.1.1
pymupdf==1.24.5
python-docx==1.1.2
openpyxl==3.1.4
python-jose==3.3.0
httpx==0.27.0
